refactor : Rename trip crew_base parameter to trip_base
Trip objects are based in a given airport which is called the trip_base.
CrewMember objects are also based in their corresponding crew_base.
Some CrewMember might be assigned to a trip starting in another base and
that is why Trip and CrewMember bases are different and their param
should be kept separte and properly identified.